Ajustes de tipo de dato y uso de ID nulo

+++ src/main/java/com/example/demo/client/PetStoreClient.java
@@ -18,7 +18,7 @@ public class PetStoreClient {
                 .build();
     }
 
-    public Pet getPetById(Integer id) {
+    public Pet getPetById(Long id) {
         return restClient.get()
                 .uri("/pet/{id}", id)
                 .retrieve()
