+++ src/main/java/com/example/demo/dto/PetResponseDto.java
@@ -1,8 +1,7 @@
 package com.example.demo.dto;
 
 public record PetResponseDto(
-        Integer id,
-        String name,
-        String status
-) {
+                Long id,
+                String name,
+                String status) {
 }
